Reorganized and cleaned unused files
