--- app/src/pages/Cashflow.actions.js
@@ -1,9 +1,0 @@
-export const retrieveProjects = {
-  types: ['LOADING_PROJECTS', 'SUCCESS_PROJECTS', 'FAILED_PROJECTS'],
-  payload: {
-    client: 'sanbaseClient',
-    request: {
-      url: `/cashflow`
-    }
-  }
-}
